Split out uploader implementation by copy file type
